feat: Support multiple namespaces on a XML Document
Fixes: #266
