+++ runtime/modules/xml/core/src/test/resources/inspect/multiple-namespaces-schemaset.xml
@@ -1,0 +1,41 @@
+<atlas:SchemaSet xmlns:atlas="http://atlasmap.io/xml/schemaset/v2" xmlns:xs="http://www.w3.org/2001/XMLSchema"
+  xmlns:first="io.atlasmap.xml.test:First" xmlns:second="io.atlasmap.xml.test:Second" >
+
+  <xs:schema targetNamespace="io.atlasmap.xml.test:Root" xmlns="io.atlasmap.xml.test:Root">
+
+    <xs:element name="RootDocument">
+      <xs:complexType>
+        <xs:sequence>
+          <xs:element name="Name" type="xs:string" />
+          <xs:element name="Value" type="xs:string" />
+          <xs:element ref="first:FirstElement" />
+          <xs:element ref="second:SecondElement" />
+        </xs:sequence>
+      </xs:complexType>
+    </xs:element>
+  </xs:schema>
+    
+  <atlas:AdditionalSchemas>
+    <xs:schema targetNamespace="io.atlasmap.xml.test:First">
+      <xs:element name="FirstElement">
+        <xs:complexType>
+          <xs:sequence>
+            <xs:element name="Name" type="xs:string" />
+            <xs:element name="Value" type="xs:string" />
+          </xs:sequence>
+        </xs:complexType>
+      </xs:element>
+    </xs:schema>
+    <xs:schema targetNamespace="io.atlasmap.xml.test:Second">
+      <xs:element name="SecondElement">
+        <xs:complexType>
+          <xs:sequence>
+            <xs:element name="Name" type="xs:string" />
+            <xs:element name="Value" type="xs:string" />
+          </xs:sequence>
+        </xs:complexType>
+      </xs:element>
+    </xs:schema>
+  </atlas:AdditionalSchemas>
+
+</atlas:SchemaSet>
